+++ src/main/scala/behaviors/play/BlackjackPlay.scala
@@ -9,8 +9,8 @@ import cards.classes.Suit._
 import cards.classes.hand.Hand
 import cards.classes.actions.{ Action, BlackjackAction }
 import cards.classes.actions.BlackjackAction._
-import cards.classes.options.BlackjackOptions
-import cards.classes.options.DealerHitLimit._
+import cards.classes.options.blackjack.BlackjackOptions
+import cards.classes.options.blackjack.DealerHitLimit._
 
 trait BlackjackPlay {
   type EVAL <: BlackjackHandEvaluation 
@@ -162,7 +162,6 @@ trait BlackjackPlay {
       history = game.history ++ newHistory, 
       dealerHand = game.dealerHand.copy(hand = newDealerCards))
 
-    // TODO: test
     // if dealer's 21 or busted or is Standing, then game is over and bets should be settled
     val gameOver: Boolean = eval(newDealerCards) >= 21 || action == Stand || nextState.history.reverse.head.action == ShowCards
     gameOver match {
@@ -206,8 +205,8 @@ trait BlackjackPlay {
     // player's turn: based on player's cards and dealers face up card, decide which action to take
     // looking at cards in reverse order so aces are at head, and changed to list in order to pattern match on head :: tail
     val highestRank: Rank = game.currentCards().sorted.reverse.head.rank
-    val totalScore: Int = eval(game.currentCards())
-    val tailScore: Int = eval(game.currentCards().tail)
+    val totalScore: Long = eval(game.currentCards())
+    val tailScore: Long = eval(game.currentCards().tail)
     // val dealerFaceUpRank: Rank = game.dealerHand.hand.head.rank
     val dealerFaceUpRank: Rank = game.dealerHand.hand.tail.head.rank
     val surrenderOffered: Boolean = game.options.allowSurrender
