+++ src/test/scala/behaviors/evaluation/ThirtyOneHandEvaluationSpec.scala
@@ -17,61 +17,61 @@ class ThirtyOneHandEvaluationSpec extends AnyFlatSpec {
 
   "ThirtyOneHandEvaluation" should "evaluate an empty hand as 0" in {
     val cards: Seq[Card] = Nil
-    val result: Int = module.eval(cards)
+    val result: Long = module.eval(cards)
     result should equal (0)
   }
 
   it should "evaluate a single Two as 2" in {
     val cards: Seq[Card] = Seq(Card(Two, Hearts))
-    val result: Int = module.eval(cards)
+    val result: Long = module.eval(cards)
     result should equal (2)
   } 
   
   it should "evaluate a single Ace as 11" in {
     val cards: Seq[Card] = Seq(Card(Ace, Diamonds))
-    val result: Int = module.eval(cards)
+    val result: Long = module.eval(cards)
     result should equal (11)
   } 
   
   it should "evaluate a Jack and Ace of different suits as 11" in {
     val cards: Seq[Card] = Seq(Card(Ace, Diamonds), Card(Jack, Clubs))
-    val result: Int = module.eval(cards)
+    val result: Long = module.eval(cards)
     result should equal (11)
   } 
 
   it should "evaluate a Jack and Ace of same suits as 21" in {
     val cards: Seq[Card] = Seq(Card(Ace, Spades), Card(Jack, Spades))
-    val result: Int = module.eval(cards)
+    val result: Long = module.eval(cards)
     result should equal (21)
   } 
 
   it should "evaluate a Ten, Jack and Ace of same suits as 32" in {
     val cards: Seq[Card] = Seq(Card(Ten, Spades), Card(Ace, Spades), Card(Jack, Spades))
-    val result: Int = module.eval(cards)
+    val result: Long = module.eval(cards)
     result should equal (32)
   } 
   
   it should "evaluate two Jacks of different suits as 10" in {
     val cards: Seq[Card] = Seq(Card(Jack, Hearts), Card(Jack, Spades))
-    val result: Int = module.eval(cards)
+    val result: Long = module.eval(cards)
     result should equal (10)
   } 
 
   it should "evaluate Jack, Queen, and King of same suits as 30" in {
     val cards: Seq[Card] = Seq(Card(Jack, Hearts), Card(Queen, Hearts), Card(King, Hearts))
-    val result: Int = module.eval(cards)
+    val result: Long = module.eval(cards)
     result should equal (30)
   } 
 
   it should "evaluate 3 Jacks from different suits as 31 (simulating 30.5)" in {
     val cards: Seq[Card] = Seq(Card(Jack, Hearts), Card(Jack, Clubs), Card(Jack, Diamonds))
-    val result: Int = module.eval(cards)
+    val result: Long = module.eval(cards)
     result should equal (31)
   } 
 
   it should "evaluate 3 Twos from different suits as 31 (simulating 30.5)" in {
     val cards: Seq[Card] = Seq(Card(Two, Hearts), Card(Two, Clubs), Card(Two, Diamonds))
-    val result: Int = module.eval(cards)
+    val result: Long = module.eval(cards)
     result should equal (31)
   }
